🚚 Rename several .yaml sources
