+++ src/pages/carnival/faq.tsx
@@ -5,14 +5,17 @@ import Layout from "components/Layout"
 import Header from "components/Header"
 
 export default function Index() {
-    const Main = tw.div`grid grid-cols-2`
-    const Sidebar = tw.side``
+    const Main = tw.main`relative`
+    const Sidebar = tw.aside`fixed`
+    const Content = tw.div`grid grid-cols-2`
 
     return (
         <Layout hasPadding isTop={false} title="有问必答">
             <Header category="carnival" titleId={2} />
-            <Sidebar></Sidebar>
-            <Main></Main>
+            <Main>
+                <Sidebar></Sidebar>
+                <Content></Content>
+            </Main>
         </Layout>
     )
 }
